Convert invalid signature test cases from Mikuli specific to generic
